Feat: Criada entidade delivery, alterado tipo do unit_price, realizada migration e atualização do context

+++ DevInSales/Models/order_product.cs
@@ -8,7 +8,7 @@ public class order_product
     public int id { get; set; }
     public int order_id { get; set; }
     public int product_id { get; set; }
-    public decimal unit_price { get; set; }
+    public float unit_price { get; set; }
     public int amount { get; set; }
     
 }
